refactor: Consolidate transaction and health check logic into dedicated modules with improved structure and utility functions
