+++ transaction-debugger/src/utils/mod.rs
@@ -1,2 +1,7 @@
 pub mod error_handling;
 pub mod validation;
+pub mod system;
+pub mod rpc;
+pub mod protobuf;
+pub mod time;
+pub mod transaction;
