+++ transaction-debugger/src/utils/protobuf.rs
@@ -1,0 +1,23 @@
+//! Protobuf utility functions for performing Protobuf-related health checks.
+//!
+//! This module provides functionality for testing protobuf message generation and serialization.
+
+/// Tests protobuf message generation and serialization.
+///
+/// Validates that gRPC protobuf message structures can be instantiated
+/// correctly, ensuring the protobuf definitions are properly compiled
+/// and accessible.
+///
+/// # Returns
+///
+/// `true` if protobuf generation is functional, `false` otherwise
+pub async fn check_protobuf_generation() -> bool {
+    use crate::grpc::transaction::debugger::DebugRequest;
+
+    let _test_request = DebugRequest {
+        signature: "test".to_string(),
+        rpc_url: "test".to_string(),
+    };
+
+    true
+}
